ia_coach/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.9.2
pydantic-settings==2.6.1
httpx==0.27.2
redis==5.2.0
python-dotenv==1.0.1
structlog==24.4.0
python-jose[cryptography]==3.3.0
openai==1.57.0
prometheus-fastapi-instrumentator==7.0.0
